requirements.txt
bcrypt==3.2.0
blinker==1.4
Bootstrap-Flask==1.7.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
dnspython==2.1.0
docopt==0.6.2
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
iteration-utilities==0.11.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pipreqs==0.4.11
pycparser==2.20
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
stripe==2.60.0
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
yarg==0.1.9
